requirements.txt
PyQt5==5.15.10
SpeechRecognition>=3.14.0
standard-aifc; python_version >= "3.13"
pyttsx3==2.90
pyaudio==0.2.14
anthropic>=0.40.0
python-dotenv==1.0.1
